+++ src/main/java/com/teammatching/demo/application/UserAccount/LoginService.java
@@ -1,0 +1,44 @@
+package com.teammatching.demo.application.UserAccount;
+
+import com.teammatching.demo.domain.dto.Principal;
+import com.teammatching.demo.domain.entity.UserAccount;
+import com.teammatching.demo.domain.repository.UserAccountRepository;
+import com.teammatching.demo.global.exception.NotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class LoginService implements UserDetailsService {
+
+    private final UserAccountRepository userAccountRepository;
+
+    /**
+     * Locates the user based on the username. In the actual implementation, the search
+     * may possibly be case sensitive, or case insensitive depending on how the
+     * implementation instance is configured. In this case, the <code>UserDetails</code>
+     * object that comes back may have a username that is of a different case than what
+     * was actually requested..
+     *
+     * @param userId the username identifying the user whose data is required.
+     * @return a fully populated user record (never <code>null</code>)
+     * @throws UsernameNotFoundException if the user could not be found or the user has no
+     *                                   GrantedAuthority
+     */
+    @Override
+    public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
+        UserAccount userAccount = userAccountRepository.findById(userId)
+                .orElseThrow(NotFoundException.UserAccount::new);
+
+        return Principal.builder()
+                .userId(userAccount.getUserId())
+                .userPassword(userAccount.getUserPassword())
+                .email(userAccount.getEmail())
+                .nickname(userAccount.getNickname())
+                .memo(userAccount.getMemo())
+                .build();
+    }
+}
