+++ src/main/java/com/teammatching/demo/global/customResponse/ResponseResult.java
@@ -1,0 +1,18 @@
+package com.teammatching.demo.global.customResponse;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Schema(hidden = true)
+@Getter
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@AllArgsConstructor
+@Builder
+public class ResponseResult<T> {
+    private int resultCode;
+    private String resultMessage;
+    private T resultData;
+}
