+++ src/main/java/com/teammatching/demo/global/exception/NullCheckException.java
@@ -1,0 +1,15 @@
+package com.teammatching.demo.global.exception;
+
+public class NullCheckException extends RuntimeException{
+    /**
+     * Constructs a new runtime exception with the specified detail message.
+     * The cause is not initialized, and may subsequently be initialized by a
+     * call to {@link #initCause}.
+     *
+     * @param message the detail message. The detail message is saved for
+     *                later retrieval by the {@link #getMessage()} method.
+     */
+    public NullCheckException(String message) {
+        super(message);
+    }
+}
