+++ src/main/java/com/teammatching/demo/global/exception/NullTokenException.java
@@ -1,0 +1,4 @@
+package com.teammatching.demo.global.exception;
+
+public class NullTokenException extends RuntimeException{
+}
