+++ src/main/java/com/teammatching/demo/global/exception/TeamJoinException.java
@@ -1,0 +1,12 @@
+package com.teammatching.demo.global.exception;
+
+public class TeamJoinException extends RuntimeException {
+    public static class AlreadyApplying extends TeamJoinException {
+    }
+
+    public static class AlreadyJoined extends TeamJoinException {
+    }
+
+    public static class FullCapacity extends TeamJoinException {
+    }
+}
