+++ src/main/java/com/teammatching/demo/global/util/PasswordUtil.java
@@ -1,0 +1,20 @@
+package com.teammatching.demo.global.util;
+
+import java.util.Random;
+
+public class PasswordUtil {
+
+    public static String generateRandomPassword() {
+        int length = 8;
+        String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()-_=+";
+        StringBuilder password = new StringBuilder();
+
+        Random random = new Random();
+        for (int i = 0; i < length; i++) {
+            int index = random.nextInt(characters.length());
+            password.append(characters.charAt(index));
+        }
+
+        return password.toString();
+    }
+}
