--- src/main/java/com/teammatching/demo/result/ResponseResult.java
@@ -1,18 +1,0 @@
-package com.teammatching.demo.result;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Schema(hidden = true)
-@Getter
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@AllArgsConstructor
-@Builder
-public class ResponseResult<T> {
-    private int resultCode;
-    private String resultMessage;
-    private T resultData;
-}
