--- src/main/java/com/teammatching/demo/result/exception/AlreadyExistsException.java
@@ -1,15 +1,0 @@
-package com.teammatching.demo.result.exception;
-
-public class AlreadyExistsException extends RuntimeException{
-    /**
-     * Constructs a new runtime exception with the specified detail message.
-     * The cause is not initialized, and may subsequently be initialized by a
-     * call to {@link #initCause}.
-     *
-     * @param message the detail message. The detail message is saved for
-     *                later retrieval by the {@link #getMessage()} method.
-     */
-    public AlreadyExistsException(String message) {
-        super(message);
-    }
-}
