--- src/main/java/com/teammatching/demo/result/exception/NullTokenException.java
@@ -1,4 +1,0 @@
-package com.teammatching.demo.result.exception;
-
-public class NullTokenException extends RuntimeException{
-}
