--- src/main/java/com/teammatching/demo/result/exception/TeamJoinException.java
@@ -1,12 +1,0 @@
-package com.teammatching.demo.result.exception;
-
-public class TeamJoinException extends RuntimeException {
-    public static class AlreadyApplying extends TeamJoinException {
-    }
-
-    public static class AlreadyJoined extends TeamJoinException {
-    }
-
-    public static class FullCapacity extends TeamJoinException {
-    }
-}
