--- src/main/java/com/teammatching/demo/web/filter/login/CustomJsonUsernamePasswordAuthenticationFilter.java
@@ -1,74 +1,0 @@
-package com.teammatching.demo.web.filter.login;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.security.authentication.AuthenticationServiceException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.util.StreamUtils;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.Map;
-
-public class CustomJsonUsernamePasswordAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
-
-    private static final String DEFAULT_LOGIN_REQUEST_URL = "/auth/login";
-    private static final String HTTP_METHOD = "POST";
-    private static final String CONTENT_TYPE = "application/json";
-    private static final String USERNAME_KEY = "userId";
-    private static final String PASSWORD_KEY = "userPassword";
-    private static final AntPathRequestMatcher DEFAULT_LOGIN_PATH_REQUEST_MATCHER =
-            new AntPathRequestMatcher(DEFAULT_LOGIN_REQUEST_URL, HTTP_METHOD);
-
-    private final ObjectMapper objectMapper;
-
-    public CustomJsonUsernamePasswordAuthenticationFilter(ObjectMapper objectMapper) {
-        super(DEFAULT_LOGIN_PATH_REQUEST_MATCHER);
-        this.objectMapper = objectMapper;
-    }
-
-    /**
-     * Performs actual authentication.
-     * <p>
-     * The implementation should do one of the following:
-     * <ol>
-     * <li>Return a populated authentication token for the authenticated user, indicating
-     * successful authentication</li>
-     * <li>Return null, indicating that the authentication process is still in progress.
-     * Before returning, the implementation should perform any additional work required to
-     * complete the process.</li>
-     * <li>Throw an <tt>AuthenticationException</tt> if the authentication process
-     * fails</li>
-     * </ol>
-     *
-     * @param request  from which to extract parameters and perform the authentication
-     * @param response the response, which may be needed if the implementation has to do a
-     *                 redirect as part of a multi-stage authentication process (such as OpenID).
-     * @return the authenticated user token, or null if authentication is incomplete.
-     * @throws AuthenticationException if authentication fails.
-     */
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException, IOException, ServletException {
-
-        if (request.getContentType() == null || !request.getContentType().equals(CONTENT_TYPE)) {
-            throw new AuthenticationServiceException("Authentication Content-Type not supported: " + request.getContentType());
-        }
-
-        String messageBody = StreamUtils.copyToString(request.getInputStream(), StandardCharsets.UTF_8);
-
-        Map<String, String> usernamePasswordMap = objectMapper.readValue(messageBody, Map.class);
-
-        String userId = usernamePasswordMap.get(USERNAME_KEY);
-        String userPassword = usernamePasswordMap.get(PASSWORD_KEY);
-
-        UsernamePasswordAuthenticationToken authRequest = new UsernamePasswordAuthenticationToken(userId, userPassword);
-
-        return this.getAuthenticationManager().authenticate(authRequest);
-    }
-}
